--- Memoria.Launcher/Memoria/UiCheckBox.cs
@@ -1,8 +1,0 @@
-using System.Windows.Controls;
-
-namespace Memoria.Launcher
-{
-    public class UiCheckBox : CheckBox
-    {
-    }
-}
