--- Memoria.Launcher/Memoria/UiComboBoxFactory.cs
@@ -1,10 +1,0 @@
-namespace Memoria.Launcher
-{
-    public static class UiComboBoxFactory
-    {
-        public static UiComboBox Create()
-        {
-            return new UiComboBox();
-        }
-    }
-}
