--- Memoria.Launcher/Memoria/UiRadioButton.cs
@@ -1,8 +1,0 @@
-using System.Windows.Controls;
-
-namespace Memoria.Launcher
-{
-    public class UiRadioButton : RadioButton
-    {
-    }
-}
