--- Memoria.Launcher/Memoria/UiSliderFactory.cs
@@ -1,16 +1,0 @@
-﻿using System.Windows.Data;
-
-namespace Memoria.Launcher
-{
-    public static class UiSliderFactory
-    {
-        public static UiSlider Create(int value)
-        {
-            return new UiSlider
-            {
-                Value = value
-            };
-        }
-    }
-}
-
