--- Memoria.Launcher/Memoria/UiTextBlock.cs
@@ -1,8 +1,0 @@
-using System.Windows.Controls;
-
-namespace Memoria.Launcher
-{
-    public class UiTextBlock : TextBlock
-    {
-    }
-}
