--- Memoria.Launcher/Memoria/UiTextBlockFactory.cs
@@ -1,14 +1,0 @@
-using System;
-
-namespace Memoria.Launcher
-{
-    public static class UiTextBlockFactory
-    {
-        public static UiTextBlock Create(String text)
-        {
-            UiTextBlock textBlock = new UiTextBlock { Text = text };
-
-            return textBlock;
-        }
-    }
-}
