--- Memoria.Launcher/Memoria/UiTextBox.cs
@@ -1,9 +1,0 @@
-﻿using System;
-using System.Windows.Controls;
-
-namespace Memoria.Launcher
-{
-    public class UiTextBox : TextBox
-    {
-    }
-}
